Alpha

+++ scripts/startup_maintenance.php
@@ -69,8 +69,14 @@ if (defined('SETTINGS_SYSTEM_DEBMES_PATH') && SETTINGS_SYSTEM_DEBMES_PATH != '')
 }
 
 DebMes("Checking log files.", 'maintenance');
+if (!isset($files) || !is_array($files)) {
+    $files = array();
+}
 getDirTree($path, $files);
 foreach ($files as $file) {
+    if (empty($file['FILENAME'])) {
+        continue;
+    }
     if (filemtime($file['FILENAME']) < time() - LOG_FILES_EXPIRE * 24 * 60 * 60) {
         DebMes("Removing log file " . $file['FILENAME'], 'maintenance');
         unlink($file['FILENAME']);
@@ -98,62 +104,66 @@ if ($full_backup) {
         copyTree(ROOT . 'cms/' . $d, $target_dir . '/cms/' . $d, 1);
     }
 
-    if (defined('PATH_TO_MYSQLDUMP'))
-        $mysqlDumpPath = PATH_TO_MYSQLDUMP;
-
-    if ($mysqlDumpPath == '') {
-        if (substr(php_uname(), 0, 7) == "Windows")
-            $mysqlDumpPath = SERVER_ROOT . "/server/mysql/bin/mysqldump";
-        else
-            $mysqlDumpPath = "/usr/bin/mysqldump";
+    $target_file = $target_dir . "/" . DB_NAME . ".sql";
+    DebMes("Backing up database " . DB_NAME . ' to ' . $target_file, 'maintenance');
+    if (SQLMakeDBDump($target_file)) {
+        DebMes("Backup done.", 'maintenance');
+    } else {
+        DebMes("Backup error.", 'maintenance');
     }
-
-    $mysqlDumpParam = " -h " . DB_HOST . " --user=\"" . DB_USER . "\" --password=\"" . DB_PASSWORD . "\"";
-    $mysqlDumpParam .= " --no-create-db --add-drop-table --databases " . DB_NAME;
-    $mysqlDumpParam .= " > " . $target_dir . "/" . DB_NAME . ".sql";
-
-    DebMes("Backing up database " . DB_NAME . ' to ' . $target_dir . "/" . DB_NAME . ".sql", 'maintenance');
-    exec($mysqlDumpPath . $mysqlDumpParam);
-    DebMes("Backup done.", 'maintenance');
     echo "OK\n";
 }
 
 
-// removing old files from cms/saverestore
-DebMes("Checking cms/saverestore files.", 'maintenance');
-if (is_dir(ROOT . 'cms/saverestore')) {
-    $files = scandir(ROOT . 'cms/saverestore');
-    foreach ($files as $file) {
-        $path = ROOT . 'cms/saverestore/' . $file;
-        if (is_file($path)
-            && (preg_match('/\.tgz$/', $file) || preg_match('/\.tar\.gz$/', $file) || preg_match('/\.zip\.gz$/', $file))
-            && filemtime($path) < time() - BACKUP_FILES_EXPIRE * 24 * 60 * 60
-        ) {
-            echo("Removing $path");
-            DebMes("Removing $path.", 'maintenance');
-            @unlink($path);
+if (!isset($run_from_start) || $run_from_start == 0) {
+    // removing old files from cms/saverestore
+    DebMes("Checking cms/saverestore files.", 'maintenance');
+    if (is_dir(ROOT . 'cms/saverestore')) {
+        $files = scandir(ROOT . 'cms/saverestore');
+        foreach ($files as $file) {
+            $path = ROOT . 'cms/saverestore/' . $file;
+            if (is_file($path)
+                && (preg_match('/\.tgz$/', $file) || preg_match('/\.tar\.gz$/', $file) || preg_match('/\.zip\.gz$/', $file))
+                && filemtime($path) < time() - BACKUP_FILES_EXPIRE * 24 * 60 * 60
+            ) {
+                echo("Removing $path");
+                DebMes("Removing $path.", 'maintenance');
+                @unlink($path);
+            }
         }
+    } else {
+        DebMes(ROOT . 'cms/saverestore - NOT FOUND', 'maintenance');
     }
-} else {
-    DebMes(ROOT . 'cms/saverestore - NOT FOUND', 'maintenance');
-}
 // removing old backus
-DebMes('Checking old backups.', 'maintenance');
-if (is_dir($backups_dir)) {
-    $backups = scandir($backups_dir);
-    foreach ($backups as $file) {
-        if ($file == '.' || $file == '..') continue;
-        $path = $backups_dir . '/' . $file;
-        if (is_dir($path) && filemtime($path) < time() - BACKUP_FILES_EXPIRE * 24 * 60 * 60) {
-            echo("Removing $path");
-            DebMes("Removing $path.", 'maintenance');
-            removeTree($path);
+    DebMes('Checking old backups.', 'maintenance');
+    if (is_dir($backups_dir)) {
+        $backups = scandir($backups_dir);
+        foreach ($backups as $file) {
+            if ($file == '.' || $file == '..') continue;
+            $path = $backups_dir . '/' . $file;
+            if (is_dir($path) && filemtime($path) < time() - BACKUP_FILES_EXPIRE * 24 * 60 * 60) {
+                echo("Removing $path");
+                DebMes("Removing $path.", 'maintenance');
+                removeTree($path);
+            }
         }
+        echo "OK";
+    } else {
+        DebMes($backups_dir . ' - NOT FOUND', 'maintenance');
+        echo $backups_dir . " not found";
     }
-    echo "OK";
 } else {
-    DebMes($backups_dir . ' - NOT FOUND', 'maintenance');
-    echo $backups_dir . " not found";
+    // RUN ON START ONLY
+    if (time() >= getGlobal('ThisComputer.started_time')) {
+        DebMes("Incorrect date on start, fixing.", 'maintenance');
+        SQLExec("DELETE FROM events WHERE ADDED > NOW()");
+        SQLExec("DELETE FROM phistory WHERE ADDED > NOW()");
+        SQLExec("DELETE FROM history WHERE ADDED > NOW()");
+        SQLExec("DELETE FROM shouts WHERE ADDED > NOW()");
+        SQLExec("DELETE FROM jobs WHERE PROCESSED = 1");
+        SQLExec("DELETE FROM history WHERE (TO_DAYS(NOW()) - TO_DAYS(ADDED)) >= 5");
+    }
+    setGlobal('ThisComputer.started_time', time());
 }
 
 
@@ -162,7 +172,10 @@ DebMes('Checking database tables.', 'maintenance');
 $tables = SQLSelect("SHOW TABLES FROM `" . DB_NAME . "`");
 $total = count($tables);
 for ($i = 0; $i < $total; $i++) {
-    $table = $tables[$i]['Tables_in_' . DB_NAME];
+    $table = $tables[$i]['Tables_in_' . DB_NAME] ?? null;
+    if (!$table) {
+        continue;
+    }
     echo 'Checking table [' . $table . '] ...';
     if ($result = SQLExec("CHECK TABLE " . $table . ";")) {
         echo "OK\n";
@@ -174,17 +187,6 @@ for ($i = 0; $i < $total; $i++) {
     }
 }
 
-setGlobal('ThisComputer.started_time', time());
-if (time() >= getGlobal('ThisComputer.started_time')) {
-    DebMes("Incorrect date on start, fixing.", 'maintenance');
-    SQLExec("DELETE FROM events WHERE ADDED > NOW()");
-    SQLExec("DELETE FROM phistory WHERE ADDED > NOW()");
-    SQLExec("DELETE FROM history WHERE ADDED > NOW()");
-    SQLExec("DELETE FROM shouts WHERE ADDED > NOW()");
-    SQLExec("DELETE FROM jobs WHERE PROCESSED = 1");
-    SQLExec("DELETE FROM history WHERE (TO_DAYS(NOW()) - TO_DAYS(ADDED)) >= 5");
-}
-
 
 // removing incorrect pvalues
 DebMes("Checking for incorrect pvalues.", 'maintenance');
@@ -193,9 +195,12 @@ $data = SQLSelect($sqlQuery);
 $total = count($data);
 $found_pvalues = array();
 for ($i = 0; $i < $total; $i++) {
-    if (!$data[$i]['PROP_ID'] || !$data[$i]['OBJ_ID']) {
-        echo "Removing incorrect property value: " . $data[$i]['PROPERTY_NAME'] . PHP_EOL;
-        DebMes("Removing incorrect property value: " . $data[$i]['PROPERTY_NAME'], 'maintenance');
+    $propId = $data[$i]['PROP_ID'] ?? null;
+    $objId = $data[$i]['OBJ_ID'] ?? null;
+    if (!$propId || !$objId) {
+        $propertyName = $data[$i]['PROPERTY_NAME'] ?? '';
+        echo "Removing incorrect property value: " . $propertyName . PHP_EOL;
+        DebMes("Removing incorrect property value: " . $propertyName, 'maintenance');
         SQLExec("DELETE FROM phistory WHERE VALUE_ID=" . $data[$i]['ID']);
         SQLExec("DELETE FROM pvalues WHERE ID=" . $data[$i]['ID']);
     } else {
@@ -220,7 +225,7 @@ $total = count($data);
 
 for ($i = 0; $i < $total; $i++) {
     $objectProperty = $data[$i]['OBJECT_TITLE'] . "." . $data[$i]['PROPERTY_TITLE'];
-    if ($data[$i]['PROPERTY_NAME']) {
+    if (!empty($data[$i]['PROPERTY_NAME'])) {
         echo "Incorrect: " . $data[$i]['PROPERTY_NAME'] . " should be $objectProperty" . PHP_EOL;
         DebMes("Incorrect: " . $data[$i]['PROPERTY_NAME'] . " should be $objectProperty", 'maintenance');
     } else {
@@ -233,9 +238,7 @@ for ($i = 0; $i < $total; $i++) {
                  WHERE ID = '" . $data[$i]['ID'] . "'";
 
     $rec = SQLSelectOne($sqlQuery);
-
     $rec['PROPERTY_NAME'] = $data[$i]['OBJECT_TITLE'] . "." . $data[$i]['PROPERTY_TITLE'];
-
     SQLUpdate('pvalues', $rec);
 }
 
@@ -302,6 +305,9 @@ for ($i = 0; $i < $total; $i++) {
         $class_id = $object_rec['CLASS_ID'];
         $class_property = array();
         $parent_props = $cls_module->getParentProperties($class_id, '', true);
+        if (!is_array($parent_props)) {
+            $parent_props = array();
+        }
         foreach ($parent_props as $class_prop) {
             if ($class_prop['TITLE'] == $prop_title) {
                 $class_property = $class_prop;
@@ -311,12 +317,16 @@ for ($i = 0; $i < $total; $i++) {
             DebMes('Fixing ' . json_encode($properties[$i]), 'maintenance');
             $object_pvalue = SQLSelectOne("SELECT * FROM pvalues WHERE PROPERTY_ID=" . $properties[$i]['ID'] . " AND OBJECT_ID=" . $properties[$i]['OBJECT_ID']);
             $class_pvalue = SQLSelectOne("SELECT * FROM pvalues WHERE PROPERTY_ID=" . $class_property['ID'] . " AND OBJECT_ID=" . $properties[$i]['OBJECT_ID']);
-            if (!$class_pvalue['ID']) {
-                $object_pvalue['PROPERTY_ID'] = $class_property['ID'];
-                SQLUpdate('pvalues', $object_pvalue);
+            if (empty($class_pvalue['ID'])) {
+                if (!empty($object_pvalue['ID'])) {
+                    $object_pvalue['PROPERTY_ID'] = $class_property['ID'];
+                    SQLUpdate('pvalues', $object_pvalue);
+                }
             } else {
-                SQLExec("DELETE FROM phistory WHERE VALUE_ID=" . $object_pvalue['ID']);
-                SQLExec("DELETE FROM pvalues WHERE ID=" . $object_pvalue['ID']);
+                if (!empty($object_pvalue['ID'])) {
+                    SQLExec("DELETE FROM phistory WHERE VALUE_ID=" . $object_pvalue['ID']);
+                    SQLExec("DELETE FROM pvalues WHERE ID=" . $object_pvalue['ID']);
+                }
             }
             SQLExec("DELETE FROM properties WHERE ID=" . $properties[$i]['ID']);
             $problems_found++;
@@ -357,6 +367,7 @@ $mkt->marketRequest('op=news');
 
 include_once DIR_MODULES . 'saverestore/saverestore.class.php';
 $sv = new saverestore();
+$out = array();
 $sv->admin($out);
 
 DebMes("Maintenance complete.", 'maintenance');
